requirements.txt
-i https://pypi.org/simple
django==2.1
djangorestframework-simplejwt==3.2.3
djangorestframework==3.8.2
git+https://github.com/chibisov/drf-extensions@9cc953bb8da6a0e08ce03f9af5c41b072aa37ecd#egg=drf-extensions
pyjwt==1.6.4
python-dotenv==0.9.1
pytz==2018.5
